Revert "gh-528: Split configure file in readable chunks"
This reverts commit 677aaf2d330225c950009e69fd4c3cced05f2418.

This is conflicting with the work in the feature/database-schema branch;
in that brach I had to touch _every_ endpoint to ensure the correct
locking of the database schema.

We can re-apply this trick once that branch is merged in.
