--- restapi/configure_weaviate_keys_helpers.go
@@ -1,67 +1,0 @@
-/*                          _       _
- *__      _____  __ ___   ___  __ _| |_ ___
- *\ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
- * \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
- *  \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
- *
- * Copyright © 2016 - 2018 Weaviate. All rights reserved.
- * LICENSE: https://github.com/creativesoftwarefdn/weaviate/blob/develop/LICENSE.md
- * AUTHOR: Bob van Luijt (bob@kub.design)
- * See www.creativesoftwarefdn.org for details
- * Contact: @CreativeSofwFdn / bob@kub.design
- */
-
-// Package restapi with all rest API functions.
-package restapi
-
-import (
-	"context"
-
-	"github.com/go-openapi/strfmt"
-
-	"github.com/creativesoftwarefdn/weaviate/auth"
-	dbconnector "github.com/creativesoftwarefdn/weaviate/connectors"
-	"github.com/creativesoftwarefdn/weaviate/connectors/utils"
-	"github.com/creativesoftwarefdn/weaviate/models"
-)
-
-func deleteKey(ctx context.Context, databaseConnector dbconnector.DatabaseConnector, parentUUID strfmt.UUID) {
-	// Find its children
-	var allIDs []strfmt.UUID
-
-	// Get all the children to remove
-	allIDs, _ = auth.GetKeyChildrenUUIDs(ctx, databaseConnector, parentUUID, false, allIDs, 0, 0)
-
-	// Append the children to the parent UUIDs to remove all
-	allIDs = append(allIDs, parentUUID)
-
-	// Delete for every child
-	for _, keyID := range allIDs {
-		// Initialize response
-		keyResponse := models.KeyGetResponse{}
-
-		// Get the key to delete
-		dbConnector.GetKey(ctx, keyID, &keyResponse)
-
-		databaseConnector.DeleteKey(ctx, &keyResponse.Key, keyID)
-	}
-}
-
-func generateMultipleRefObject(keyIDs []strfmt.UUID) models.MultipleRef {
-	// Init the response
-	refs := models.MultipleRef{}
-
-	// Init localhost
-	url := serverConfig.GetHostAddress()
-
-	// Generate SingleRefs
-	for _, keyID := range keyIDs {
-		refs = append(refs, &models.SingleRef{
-			LocationURL:  &url,
-			NrDollarCref: keyID,
-			Type:         string(connutils.RefTypeKey),
-		})
-	}
-
-	return refs
-}
