--- restapi/configure_weaviate_things.go
@@ -1,469 +1,0 @@
-/*                          _       _
- *__      _____  __ ___   ___  __ _| |_ ___
- *\ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
- * \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
- *  \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
- *
- * Copyright © 2016 - 2018 Weaviate. All rights reserved.
- * LICENSE: https://github.com/creativesoftwarefdn/weaviate/blob/develop/LICENSE.md
- * AUTHOR: Bob van Luijt (bob@kub.design)
- * See www.creativesoftwarefdn.org for details
- * Contact: @CreativeSofwFdn / bob@kub.design
- */
-
-// Package restapi with all rest API functions.
-package restapi
-
-import (
-	"encoding/json"
-
-	jsonpatch "github.com/evanphx/json-patch"
-	middleware "github.com/go-openapi/runtime/middleware"
-	"github.com/go-openapi/strfmt"
-
-	"github.com/creativesoftwarefdn/weaviate/auth"
-	"github.com/creativesoftwarefdn/weaviate/broker"
-	"github.com/creativesoftwarefdn/weaviate/connectors/utils"
-	"github.com/creativesoftwarefdn/weaviate/models"
-	libnetwork "github.com/creativesoftwarefdn/weaviate/network"
-	"github.com/creativesoftwarefdn/weaviate/restapi/operations"
-	"github.com/creativesoftwarefdn/weaviate/restapi/operations/meta"
-	"github.com/creativesoftwarefdn/weaviate/restapi/operations/p2_p"
-	"github.com/creativesoftwarefdn/weaviate/restapi/operations/things"
-	"github.com/creativesoftwarefdn/weaviate/validation"
-)
-
-func configureAPIThings(api *operations.WeaviateAPI) {
-	api.ThingsWeaviateThingsCreateHandler = things.WeaviateThingsCreateHandlerFunc(func(params things.WeaviateThingsCreateParams, principal interface{}) middleware.Responder {
-		// Get context from request
-		ctx := params.HTTPRequest.Context()
-
-		// This is a write function, validate if allowed to write?
-		if allowed, _ := auth.ActionsAllowed(ctx, []string{"write"}, principal, dbConnector, nil); !allowed {
-			return things.NewWeaviateThingsCreateForbidden()
-		}
-
-		// Generate UUID for the new object
-		UUID := connutils.GenerateUUID()
-
-		// Convert principal to object
-		keyToken := principal.(*models.KeyTokenGetResponse)
-
-		// Validate schema given in body with the weaviate schema
-		validatedErr := validation.ValidateThingBody(params.HTTPRequest.Context(), params.Body.Thing, databaseSchema, dbConnector, serverConfig, keyToken)
-		if validatedErr != nil {
-			return things.NewWeaviateThingsCreateUnprocessableEntity().WithPayload(createErrorResponseObject(validatedErr.Error()))
-		}
-
-		// Create Key-ref-Object
-		url := serverConfig.GetHostAddress()
-		keyRef := &models.SingleRef{
-			LocationURL:  &url,
-			NrDollarCref: keyToken.KeyID,
-			Type:         string(connutils.RefTypeKey),
-		}
-
-		// Make Thing-Object
-		thing := &models.Thing{}
-		thing.Schema = params.Body.Thing.Schema
-		thing.AtClass = params.Body.Thing.AtClass
-		thing.AtContext = params.Body.Thing.AtContext
-		thing.CreationTimeUnix = connutils.NowUnix()
-		thing.LastUpdateTimeUnix = 0
-		thing.Key = keyRef
-
-		responseObject := &models.ThingGetResponse{}
-		responseObject.Thing = *thing
-		responseObject.ThingID = UUID
-
-		if params.Body.Async {
-			go dbConnector.AddThing(ctx, thing, UUID)
-			return things.NewWeaviateThingsCreateAccepted().WithPayload(responseObject)
-		} else {
-			dbConnector.AddThing(ctx, thing, UUID)
-			return things.NewWeaviateThingsCreateOK().WithPayload(responseObject)
-		}
-	})
-	api.ThingsWeaviateThingsDeleteHandler = things.WeaviateThingsDeleteHandlerFunc(func(params things.WeaviateThingsDeleteParams, principal interface{}) middleware.Responder {
-		// Initialize response
-		thingGetResponse := models.ThingGetResponse{}
-		thingGetResponse.Schema = map[string]models.JSONObject{}
-
-		// Get context from request
-		ctx := params.HTTPRequest.Context()
-
-		// Get item from database
-		errGet := dbConnector.GetThing(params.HTTPRequest.Context(), params.ThingID, &thingGetResponse)
-
-		// Save the old-thing in a variable
-		oldThing := thingGetResponse
-
-		// Not found
-		if errGet != nil {
-			return things.NewWeaviateThingsDeleteNotFound()
-		}
-
-		// This is a delete function, validate if allowed to delete?
-		if allowed, _ := auth.ActionsAllowed(ctx, []string{"delete"}, principal, dbConnector, thingGetResponse.Key.NrDollarCref); !allowed {
-			return things.NewWeaviateThingsDeleteForbidden()
-		}
-
-		// Delete the Actions
-		actionsExist := true
-		lastActionsCount := int64(0)
-		for actionsExist {
-			actions := models.ActionsListResponse{}
-			actions.Actions = []*models.ActionGetResponse{}
-			dbConnector.ListActions(ctx, params.ThingID, 50, 0, []*connutils.WhereQuery{}, &actions)
-			for _, v := range actions.Actions {
-				go dbConnector.DeleteAction(ctx, &v.Action, v.ActionID)
-			}
-
-			// Exit if total results are 0 or the total results are not lowering, then there is some kind of error
-			actionsExist = (actions.TotalResults > 0 && actions.TotalResults != lastActionsCount)
-			lastActionsCount = actions.TotalResults
-		}
-
-		thingGetResponse.LastUpdateTimeUnix = connutils.NowUnix()
-
-		// Move the current properties to the history
-		go dbConnector.MoveToHistoryThing(ctx, &oldThing.Thing, params.ThingID, true)
-
-		// Add new row as GO-routine
-		go dbConnector.DeleteThing(ctx, &thingGetResponse.Thing, params.ThingID)
-
-		// Return 'No Content'
-		return things.NewWeaviateThingsDeleteNoContent()
-	})
-	api.ThingsWeaviateThingsGetHandler = things.WeaviateThingsGetHandlerFunc(func(params things.WeaviateThingsGetParams, principal interface{}) middleware.Responder {
-		// Initialize response
-		responseObject := models.ThingGetResponse{}
-		responseObject.Schema = map[string]models.JSONObject{}
-
-		// Get context from request
-		ctx := params.HTTPRequest.Context()
-
-		// Get item from database
-		err := dbConnector.GetThing(ctx, strfmt.UUID(params.ThingID), &responseObject)
-
-		// Object is not found
-		if err != nil || responseObject.Key == nil {
-			messaging.ErrorMessage(err)
-			return things.NewWeaviateThingsGetNotFound()
-		}
-
-		// This is a read function, validate if allowed to read?
-		if allowed, _ := auth.ActionsAllowed(ctx, []string{"read"}, principal, dbConnector, responseObject.Key.NrDollarCref); !allowed {
-			return things.NewWeaviateThingsGetForbidden()
-		}
-
-		// Get is successful
-		return things.NewWeaviateThingsGetOK().WithPayload(&responseObject)
-	})
-
-	api.ThingsWeaviateThingHistoryGetHandler = things.WeaviateThingHistoryGetHandlerFunc(func(params things.WeaviateThingHistoryGetParams, principal interface{}) middleware.Responder {
-		// Initialize response
-		responseObject := models.ThingGetResponse{}
-		responseObject.Schema = map[string]models.JSONObject{}
-
-		// Get context from request
-		ctx := params.HTTPRequest.Context()
-
-		// Set UUID var for easy usage
-		UUID := strfmt.UUID(params.ThingID)
-
-		// Get item from database
-		errGet := dbConnector.GetThing(params.HTTPRequest.Context(), UUID, &responseObject)
-
-		// Init the response variables
-		historyResponse := &models.ThingGetHistoryResponse{}
-		historyResponse.PropertyHistory = []*models.ThingHistoryObject{}
-		historyResponse.ThingID = UUID
-
-		// Fill the history for these objects
-		errHist := dbConnector.HistoryThing(ctx, UUID, &historyResponse.ThingHistory)
-
-		// Check whether dont exist (both give an error) to return a not found
-		if errGet != nil && (errHist != nil || len(historyResponse.PropertyHistory) == 0) {
-			messaging.ErrorMessage(errGet)
-			messaging.ErrorMessage(errHist)
-			return things.NewWeaviateThingHistoryGetNotFound()
-		}
-
-		// This is a read function, validate if allowed to read?
-		if errHist == nil {
-			if allowed, _ := auth.ActionsAllowed(ctx, []string{"read"}, principal, dbConnector, historyResponse.Key.NrDollarCref); !allowed {
-				return things.NewWeaviateThingHistoryGetForbidden()
-			}
-		} else if errGet == nil {
-			if allowed, _ := auth.ActionsAllowed(ctx, []string{"read"}, principal, dbConnector, responseObject.Key.NrDollarCref); !allowed {
-				return things.NewWeaviateThingHistoryGetForbidden()
-			}
-		}
-
-		// Thing is deleted when we have an get error and no history error
-		historyResponse.Deleted = errGet != nil && errHist == nil && len(historyResponse.PropertyHistory) != 0
-
-		return things.NewWeaviateThingHistoryGetOK().WithPayload(historyResponse)
-	})
-
-	api.ThingsWeaviateThingsListHandler = things.WeaviateThingsListHandlerFunc(func(params things.WeaviateThingsListParams, principal interface{}) middleware.Responder {
-		// Get limit and page
-		limit := getLimit(params.MaxResults)
-		page := getPage(params.Page)
-
-		// Get user out of principal
-		keyID := principal.(*models.KeyTokenGetResponse).KeyID
-
-		// Get context from request
-		ctx := params.HTTPRequest.Context()
-
-		// This is a read function, validate if allowed to read?
-		if allowed, _ := auth.ActionsAllowed(ctx, []string{"read"}, principal, dbConnector, keyID); !allowed {
-			return things.NewWeaviateThingsListForbidden()
-		}
-
-		// Initialize response
-		thingsResponse := models.ThingsListResponse{}
-		thingsResponse.Things = []*models.ThingGetResponse{}
-
-		// List all results
-		err := dbConnector.ListThings(ctx, limit, (page-1)*limit, keyID, []*connutils.WhereQuery{}, &thingsResponse)
-
-		if err != nil {
-			messaging.ErrorMessage(err)
-		}
-
-		return things.NewWeaviateThingsListOK().WithPayload(&thingsResponse)
-	})
-	api.ThingsWeaviateThingsPatchHandler = things.WeaviateThingsPatchHandlerFunc(func(params things.WeaviateThingsPatchParams, principal interface{}) middleware.Responder {
-		// Initialize response
-		thingGetResponse := models.ThingGetResponse{}
-		thingGetResponse.Schema = map[string]models.JSONObject{}
-
-		// Get context from request
-		ctx := params.HTTPRequest.Context()
-
-		// Get and transform object
-		UUID := strfmt.UUID(params.ThingID)
-		errGet := dbConnector.GetThing(params.HTTPRequest.Context(), UUID, &thingGetResponse)
-
-		// Save the old-thing in a variable
-		oldThing := thingGetResponse
-
-		// Add update time
-		thingGetResponse.LastUpdateTimeUnix = connutils.NowUnix()
-
-		// Return error if UUID is not found.
-		if errGet != nil {
-			return things.NewWeaviateThingsPatchNotFound()
-		}
-
-		// This is a write function, validate if allowed to write?
-		if allowed, _ := auth.ActionsAllowed(ctx, []string{"write"}, principal, dbConnector, thingGetResponse.Key.NrDollarCref); !allowed {
-			return things.NewWeaviateThingsPatchForbidden()
-		}
-
-		// Get PATCH params in format RFC 6902
-		jsonBody, marshalErr := json.Marshal(params.Body)
-		patchObject, decodeErr := jsonpatch.DecodePatch([]byte(jsonBody))
-
-		if marshalErr != nil || decodeErr != nil {
-			return things.NewWeaviateThingsPatchBadRequest()
-		}
-
-		// Convert ThingGetResponse object to JSON
-		thingUpdateJSON, marshalErr := json.Marshal(thingGetResponse)
-		if marshalErr != nil {
-			return things.NewWeaviateThingsPatchBadRequest()
-		}
-
-		// Apply the patch
-		updatedJSON, applyErr := patchObject.Apply(thingUpdateJSON)
-
-		if applyErr != nil {
-			return things.NewWeaviateThingsPatchUnprocessableEntity().WithPayload(createErrorResponseObject(applyErr.Error()))
-		}
-
-		// Turn it into a Thing object
-		thing := &models.Thing{}
-		json.Unmarshal([]byte(updatedJSON), &thing)
-
-		// Convert principal to object
-		keyToken := principal.(*models.KeyTokenGetResponse)
-
-		// Validate schema made after patching with the weaviate schema
-		validatedErr := validation.ValidateThingBody(params.HTTPRequest.Context(), &thing.ThingCreate, databaseSchema, dbConnector, serverConfig, keyToken)
-		if validatedErr != nil {
-			return things.NewWeaviateThingsPatchUnprocessableEntity().WithPayload(createErrorResponseObject(validatedErr.Error()))
-		}
-
-		// Move the current properties to the history
-		go dbConnector.MoveToHistoryThing(ctx, &oldThing.Thing, UUID, false)
-
-		// Update the database
-		go dbConnector.UpdateThing(ctx, thing, UUID)
-
-		// Create return Object
-		thingGetResponse.Thing = *thing
-
-		// Returns accepted so a Go routine can process in the background
-		return things.NewWeaviateThingsPatchAccepted().WithPayload(&thingGetResponse)
-	})
-	api.ThingsWeaviateThingsPropertiesCreateHandler = things.WeaviateThingsPropertiesCreateHandlerFunc(func(params things.WeaviateThingsPropertiesCreateParams, principal interface{}) middleware.Responder {
-		return middleware.NotImplemented("operation things.WeaviateThingsPropertiesCreate has not yet been implemented")
-	})
-	api.ThingsWeaviateThingsPropertiesDeleteHandler = things.WeaviateThingsPropertiesDeleteHandlerFunc(func(params things.WeaviateThingsPropertiesDeleteParams, principal interface{}) middleware.Responder {
-		return middleware.NotImplemented("operation things.WeaviateThingsPropertiesDelete has not yet been implemented")
-	})
-	api.ThingsWeaviateThingsPropertiesUpdateHandler = things.WeaviateThingsPropertiesUpdateHandlerFunc(func(params things.WeaviateThingsPropertiesUpdateParams, principal interface{}) middleware.Responder {
-		return middleware.NotImplemented("operation things.WeaviateThingsPropertiesUpdate has not yet been implemented")
-	})
-	api.ThingsWeaviateThingsUpdateHandler = things.WeaviateThingsUpdateHandlerFunc(func(params things.WeaviateThingsUpdateParams, principal interface{}) middleware.Responder {
-		// Initialize response
-		thingGetResponse := models.ThingGetResponse{}
-		thingGetResponse.Schema = map[string]models.JSONObject{}
-
-		// Get context from request
-		ctx := params.HTTPRequest.Context()
-
-		// Get item from database
-		UUID := params.ThingID
-		errGet := dbConnector.GetThing(params.HTTPRequest.Context(), UUID, &thingGetResponse)
-
-		// Save the old-thing in a variable
-		oldThing := thingGetResponse
-
-		// If there are no results, there is an error
-		if errGet != nil {
-			// Object not found response.
-			return things.NewWeaviateThingsUpdateNotFound()
-		}
-
-		// This is a write function, validate if allowed to write?
-		if allowed, _ := auth.ActionsAllowed(ctx, []string{"write"}, principal, dbConnector, thingGetResponse.Key.NrDollarCref); !allowed {
-			return things.NewWeaviateThingsUpdateForbidden()
-		}
-
-		// Convert principal to object
-		keyToken := principal.(*models.KeyTokenGetResponse)
-
-		// Validate schema given in body with the weaviate schema
-		validatedErr := validation.ValidateThingBody(params.HTTPRequest.Context(), &params.Body.ThingCreate, databaseSchema, dbConnector, serverConfig, keyToken)
-		if validatedErr != nil {
-			return things.NewWeaviateThingsUpdateUnprocessableEntity().WithPayload(createErrorResponseObject(validatedErr.Error()))
-		}
-
-		// Move the current properties to the history
-		go dbConnector.MoveToHistoryThing(ctx, &oldThing.Thing, UUID, false)
-
-		// Update the database
-		params.Body.LastUpdateTimeUnix = connutils.NowUnix()
-		params.Body.CreationTimeUnix = thingGetResponse.CreationTimeUnix
-		params.Body.Key = thingGetResponse.Key
-		go dbConnector.UpdateThing(ctx, &params.Body.Thing, UUID)
-
-		// Create object to return
-		responseObject := &models.ThingGetResponse{}
-		responseObject.Thing = params.Body.Thing
-		responseObject.ThingID = UUID
-
-		// broadcast to MQTT
-		mqttJson, _ := json.Marshal(responseObject)
-		weaviateBroker.Publish("/things/"+string(responseObject.ThingID), string(mqttJson[:]))
-
-		// Return SUCCESS (NOTE: this is ACCEPTED, so the dbConnector.Add should have a go routine)
-		return things.NewWeaviateThingsUpdateAccepted().WithPayload(responseObject)
-	})
-	api.ThingsWeaviateThingsValidateHandler = things.WeaviateThingsValidateHandlerFunc(func(params things.WeaviateThingsValidateParams, principal interface{}) middleware.Responder {
-		// Convert principal to object
-		keyToken := principal.(*models.KeyTokenGetResponse)
-
-		// Validate schema given in body with the weaviate schema
-		validatedErr := validation.ValidateThingBody(params.HTTPRequest.Context(), params.Body, databaseSchema, dbConnector, serverConfig, keyToken)
-		if validatedErr != nil {
-			return things.NewWeaviateThingsValidateUnprocessableEntity().WithPayload(createErrorResponseObject(validatedErr.Error()))
-		}
-
-		return things.NewWeaviateThingsValidateOK()
-	})
-	api.MetaWeaviateMetaGetHandler = meta.WeaviateMetaGetHandlerFunc(func(params meta.WeaviateMetaGetParams, principal interface{}) middleware.Responder {
-		// Create response object
-		metaResponse := &models.Meta{}
-
-		// Set the response object's values
-		metaResponse.Hostname = serverConfig.GetHostAddress()
-		metaResponse.ActionsSchema = databaseSchema.ActionSchema.Schema
-		metaResponse.ThingsSchema = databaseSchema.ThingSchema.Schema
-
-		return meta.NewWeaviateMetaGetOK().WithPayload(metaResponse)
-	})
-
-	api.P2PWeaviateP2pGenesisUpdateHandler = p2_p.WeaviateP2pGenesisUpdateHandlerFunc(func(params p2_p.WeaviateP2pGenesisUpdateParams) middleware.Responder {
-		new_peers := make([]libnetwork.Peer, 0)
-
-		for _, genesis_peer := range params.Peers {
-			peer := libnetwork.Peer{
-				Id:   genesis_peer.ID,
-				Name: genesis_peer.Name,
-				URI:  genesis_peer.URI,
-			}
-
-			new_peers = append(new_peers, peer)
-		}
-
-		err := network.UpdatePeers(new_peers)
-
-		if err == nil {
-			return p2_p.NewWeaviateP2pGenesisUpdateOK()
-		} else {
-			return p2_p.NewWeaviateP2pGenesisUpdateInternalServerError()
-		}
-	})
-
-	api.P2PWeaviateP2pHealthHandler = p2_p.WeaviateP2pHealthHandlerFunc(func(params p2_p.WeaviateP2pHealthParams) middleware.Responder {
-		// For now, always just return success.
-		return middleware.NotImplemented("operation P2PWeaviateP2pHealth has not yet been implemented")
-	})
-
-	api.ThingsWeaviateThingsActionsListHandler = things.WeaviateThingsActionsListHandlerFunc(func(params things.WeaviateThingsActionsListParams, principal interface{}) middleware.Responder {
-		// Get limit and page
-		limit := getLimit(params.MaxResults)
-		page := getPage(params.Page)
-
-		// Get key-object
-		keyObject := principal.(*models.KeyTokenGetResponse)
-
-		// Get context from request
-		ctx := params.HTTPRequest.Context()
-
-		// This is a read function, validate if allowed to read?
-		if allowed, _ := auth.ActionsAllowed(ctx, []string{"read"}, principal, dbConnector, keyObject.KeyID); !allowed {
-			return things.NewWeaviateThingsActionsListForbidden()
-		}
-
-		// Initialize response
-		thingGetResponse := models.ThingGetResponse{}
-		thingGetResponse.Schema = map[string]models.JSONObject{}
-		errGet := dbConnector.GetThing(params.HTTPRequest.Context(), params.ThingID, &thingGetResponse)
-
-		// If there are no results, there is an error
-		if errGet != nil {
-			// Object not found response.
-			return things.NewWeaviateThingsActionsListNotFound()
-		}
-
-		// Initialize response
-		actionsResponse := models.ActionsListResponse{}
-		actionsResponse.Actions = []*models.ActionGetResponse{}
-
-		// List all results
-		err := dbConnector.ListActions(ctx, params.ThingID, limit, (page-1)*limit, []*connutils.WhereQuery{}, &actionsResponse)
-
-		if err != nil {
-			messaging.ErrorMessage(err)
-		}
-
-		return things.NewWeaviateThingsActionsListOK().WithPayload(&actionsResponse)
-	})
-}
